move Logo.jsx to own folder
